1. "是否強制禁止邊緣通行"設定改在各組螢幕設定檔儲存。 2. 優化執行緒安全。 3. 整理程式碼。 4. 補註解。

+++ MonitorEdgeTunnelLib/MonitorInfoManager.cpp
@@ -134,7 +134,7 @@
     }
 
     // base64½s½X
-    result = base64_encode(bytes, size);
+    result = base64_encode(bytes, static_cast<uint32_t>(size));
 
     delete[] bytes;
